Staff Type List Items have an icon now to match the concept art
