+++ Assets/Code/User Interface/Staff/StaffListItem.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 using UnityEngine.UI;
 using TMPro;
 using Code.World_Objects.Staff;
@@ -8,23 +6,14 @@ using Code.World_Objects.Staff;
 namespace Code.User_Interface {
   //Represents one selectable Staff item in the Staff List UI screen
   public class StaffListItem : WorldObjectBehaviorListItem<StaffBehavior> {
-    [Serializable]
-    public class StaffTypeIcon {
-      [Tooltip("Staff Type for this icon")]
-      public StaffType type;
-      [Tooltip("Icon for this StaffType")]
-      public Sprite sprite;
-    }
-
     [Tooltip("The element that should display the item's label string.")]
     public TMP_Text label;
     [Tooltip("The element that should display whether the staff is hired or not.")]
     public Toggle hiredToggle;
     [Tooltip("The element that should display the staff's type.")]
     public Image staffTypeIcon;
-    [Header("Customization")]
     [Tooltip("The map of StaffType to icon Sprites")]
-    public List<StaffTypeIcon> staffTypeIconMap = new List<StaffTypeIcon>();
+    public StaffTypeIconMap staffTypeIconMap;
 
     private StaffBehavior _item;
 
@@ -43,22 +32,12 @@ namespace Code.User_Interface {
     }
 
     //-------------------------------------------------------------------------
-    private Sprite GetStaffTypeIcon(StaffType type) {
-      foreach (var staffTypeIcon in staffTypeIconMap) {
-        if (staffTypeIcon.type == type) {
-          return staffTypeIcon.sprite;
-        }
-      }
-      return null;
-    }
-
-    //-------------------------------------------------------------------------
     private void UpdateUI() {
       if (this != null) {
         this.name = _item.Data.user_name;
         label.text = _item.Data.user_name;
         hiredToggle.isOn = _item.Data.IsCurrentlyHired();
-        staffTypeIcon.sprite = GetStaffTypeIcon(_item.Data.type);
+        staffTypeIcon.sprite = staffTypeIconMap.GetStaffTypeIcon(_item.Data.type);
         if (staffTypeIcon.sprite == null) {
           staffTypeIcon.enabled = false;
         }
