+++ Assets/Code/User Interface/Staff/StaffTypeIconMap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60c2c3a0e401ce241a53e1b0eafeae98
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
