+++ Assets/Prefabs/User Interface Elements/Staff View/Staff List Item.prefab
@@ -318,11 +318,7 @@ MonoBehaviour:
   label: {fileID: 3327119672017929347}
   hiredToggle: {fileID: 3327119671147883815}
   staffTypeIcon: {fileID: 3327119671664201237}
-  staffTypeIconMap:
-  - type: 0
-    sprite: {fileID: 21300000, guid: 8c9eb6ca3822bfd4f89a68de04e26b80, type: 3}
-  - type: 1
-    sprite: {fileID: 21300000, guid: 6d63dc5422207f047a0b5d6d4629767f, type: 3}
+  staffTypeIconMap: {fileID: 11400000, guid: d9937e903960c714b9056392a7b34c8a, type: 2}
 --- !u!1 &3327119671664201322
 GameObject:
   m_ObjectHideFlags: 0
@@ -566,7 +562,7 @@ MonoBehaviour:
   m_fontSizeMin: 10
   m_fontSizeMax: 36
   m_fontStyle: 0
-  m_HorizontalAlignment: 2
+  m_HorizontalAlignment: 1
   m_VerticalAlignment: 512
   m_textAlignment: 65535
   m_characterSpacing: 0
