+++ Assets/ScriptableObjects/Staff/Staff Type Icon Map.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 60c2c3a0e401ce241a53e1b0eafeae98, type: 3}
+  m_Name: Staff Type Icon Map
+  m_EditorClassIdentifier: 
+  map:
+  - type: 0
+    sprite: {fileID: 21300000, guid: 8c9eb6ca3822bfd4f89a68de04e26b80, type: 3}
+  - type: 1
+    sprite: {fileID: 21300000, guid: 6d63dc5422207f047a0b5d6d4629767f, type: 3}
