+++ Assets/ScriptableObjects/Staff/Staff Type Icon Map.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d9937e903960c714b9056392a7b34c8a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
